refactor: separar lógica de thunks de lógica de slice
