+++ store/slices/patientSlice.ts
@@ -1,8 +1,7 @@
 // slices/patientSlice.ts
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { Patient } from '@/types';
-import { TypeNewPatient } from '@/types';
-import type { RootState } from '../index';
+import { createSlice } from '@reduxjs/toolkit';
+import { createPatient, deletePatient, fetchPatients } from '@/store/thunks';
+import type { Patient } from '@/types';
 
 interface PatientState {
   list: Patient[];
@@ -15,101 +14,6 @@ const initialState: PatientState = {
   loading: false,
   error: null,
 };
-
-export const fetchPatients = createAsyncThunk<Patient[], void, { rejectValue: string }>(
-  'patients/fetchPatients',
-  async (_, thunkApi) => {
-    const state = thunkApi.getState() as RootState;
-    const token = state.auth.token;
-
-    try {
-      const response = await fetch(
-        'https://proyecto-foo-production.up.railway.app/api/Patient/pacientes',
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-
-      console.log('Respuesta recibida:', response);
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Error HTTP ${response.status}: ${errorText}`);
-      }
-
-      const data = await response.json();
-      return data;
-    } catch (error: unknown) {
-      const message = error instanceof Error ? error.message : 'Error desconocido';
-      return thunkApi.rejectWithValue(message);
-    }
-  }
-);
-
-// borro
-export const deletePatient = createAsyncThunk<number, number, { rejectValue: string }>(
-  'patients/deletePatient',
-  async (id, thunkApi) => {
-    const state = thunkApi.getState() as RootState;
-    const token = state.auth.token;
-
-    try {
-      const response = await fetch(
-        `https://proyecto-foo-production.up.railway.app/api/Patient/${id}`,
-        {
-          method: 'DELETE',
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Error al eliminar paciente: ${errorText}`);
-      }
-
-      return id;
-    } catch (error: unknown) {
-      const message = error instanceof Error ? error.message : 'Error desconocido';
-      return thunkApi.rejectWithValue(message);
-    }
-  }
-);
-
-//Crear paciente
-export const createPatient = createAsyncThunk<Patient, TypeNewPatient, { rejectValue: string }>(
-  'patients/createPatient',
-  async (newPatient, thunkApi) => {
-    const state = thunkApi.getState() as RootState;
-    const token = state.auth.token;
-
-    try {
-      console.log(newPatient);
-      const response = await fetch('https://proyecto-foo-production.up.railway.app/api/Patient', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-        body: JSON.stringify(newPatient),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Error al crear paciente: ${errorText}`);
-      }
-
-      const data = await response.json();
-      return data;
-    } catch (error: unknown) {
-      const message = error instanceof Error ? error.message : 'Error desconocido';
-      return thunkApi.rejectWithValue(message);
-    }
-  }
-);
 
 const patientSlice = createSlice({
   name: 'patients',
