+++ store/thunks/index.ts
@@ -1,0 +1,1 @@
+export { createPatient, deletePatient, fetchPatients } from './patientThunks';
